+++ server/src/decorators.ts
@@ -3,8 +3,8 @@ import { ApiExtension, ApiOperation, ApiProperty, ApiTags } from '@nestjs/swagge
 import _ from 'lodash';
 import { ADDED_IN_PREFIX, DEPRECATED_IN_PREFIX, LIFECYCLE_EXTENSION } from 'src/constants';
 import { ImmichWorker, MetadataKey } from 'src/enum';
-import { EmitEvent } from 'src/interfaces/event.interface';
 import { JobName, QueueName } from 'src/interfaces/job.interface';
+import { EmitEvent } from 'src/repositories/event.repository';
 import { setUnion } from 'src/utils/set';
 
 // PostgreSQL uses a 16-bit integer to indicate the number of bound parameters. This means that the
