+++ server/src/enum.ts
@@ -391,3 +391,10 @@ export enum DatabaseExtension {
   VECTOR = 'vector',
   VECTORS = 'vectors',
 }
+
+export enum BootstrapEventPriority {
+  // Database service should be initialized before anything else, most other services need database access
+  DatabaseService = -200,
+  // Initialise config after other bootstrap services, stop other services from using config on bootstrap
+  SystemConfig = 100,
+}
