+++ server/src/services/database.service.ts
@@ -2,8 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { Duration } from 'luxon';
 import semver from 'semver';
 import { OnEvent } from 'src/decorators';
-import { DatabaseExtension } from 'src/enum';
-import { BootstrapEventPriority } from 'src/interfaces/event.interface';
+import { BootstrapEventPriority, DatabaseExtension } from 'src/enum';
 import { DatabaseLock, EXTENSION_NAMES, VectorExtension, VectorIndex } from 'src/repositories/database.repository';
 import { BaseService } from 'src/services/base.service';
