+++ server/src/services/job.service.ts
@@ -4,7 +4,6 @@ import { OnEvent } from 'src/decorators';
 import { mapAsset } from 'src/dtos/asset-response.dto';
 import { AllJobStatusResponseDto, JobCommandDto, JobCreateDto, JobStatusDto } from 'src/dtos/job.dto';
 import { AssetType, ImmichWorker, ManualJobName } from 'src/enum';
-import { ArgOf, ArgsOf } from 'src/interfaces/event.interface';
 import {
   ConcurrentQueueName,
   JobCommand,
@@ -14,6 +13,7 @@ import {
   QueueCleanType,
   QueueName,
 } from 'src/interfaces/job.interface';
+import { ArgOf, ArgsOf } from 'src/repositories/event.repository';
 import { BaseService } from 'src/services/base.service';
 
 const asJobItem = (dto: JobCreateDto): JobItem => {
