+++ server/src/services/person.service.spec.ts
@@ -4,8 +4,8 @@ import { mapFaces, mapPerson, PersonResponseDto } from 'src/dtos/person.dto';
 import { AssetFaceEntity } from 'src/entities/asset-face.entity';
 import { CacheControl, Colorspace, ImageFormat, SourceType, SystemMetadataKey } from 'src/enum';
 import { JobName, JobStatus } from 'src/interfaces/job.interface';
-import { DetectedFaces } from 'src/interfaces/machine-learning.interface';
 import { WithoutProperty } from 'src/repositories/asset.repository';
+import { DetectedFaces } from 'src/repositories/machine-learning.repository';
 import { FaceSearchResult } from 'src/repositories/search.repository';
 import { PersonService } from 'src/services/person.service';
 import { ImmichFileResponse } from 'src/utils/file';
