+++ server/src/services/person.service.ts
@@ -40,8 +40,8 @@ import {
   JobStatus,
   QueueName,
 } from 'src/interfaces/job.interface';
-import { BoundingBox } from 'src/interfaces/machine-learning.interface';
 import { WithoutProperty } from 'src/repositories/asset.repository';
+import { BoundingBox } from 'src/repositories/machine-learning.repository';
 import { UpdateFacesData } from 'src/repositories/person.repository';
 import { BaseService } from 'src/services/base.service';
 import { CropOptions, ImageDimensions, InputDimensions } from 'src/types';
