+++ server/src/services/smart-info.service.ts
@@ -2,10 +2,10 @@ import { Injectable } from '@nestjs/common';
 import { SystemConfig } from 'src/config';
 import { OnEvent, OnJob } from 'src/decorators';
 import { ImmichWorker } from 'src/enum';
-import { ArgOf } from 'src/interfaces/event.interface';
 import { JOBS_ASSET_PAGINATION_SIZE, JobName, JobOf, JobStatus, QueueName } from 'src/interfaces/job.interface';
 import { WithoutProperty } from 'src/repositories/asset.repository';
 import { DatabaseLock } from 'src/repositories/database.repository';
+import { ArgOf } from 'src/repositories/event.repository';
 import { BaseService } from 'src/services/base.service';
 import { getAssetFiles } from 'src/utils/asset.util';
 import { getCLIPModelInfo, isSmartSearchEnabled } from 'src/utils/misc';
