+++ server/src/services/system-config.service.ts
@@ -4,7 +4,8 @@ import _ from 'lodash';
 import { defaults } from 'src/config';
 import { OnEvent } from 'src/decorators';
 import { SystemConfigDto, mapConfig } from 'src/dtos/system-config.dto';
-import { ArgOf, BootstrapEventPriority } from 'src/interfaces/event.interface';
+import { BootstrapEventPriority } from 'src/enum';
+import { ArgOf } from 'src/repositories/event.repository';
 import { BaseService } from 'src/services/base.service';
 import { clearConfigCache } from 'src/utils/config';
 import { toPlainObject } from 'src/utils/object';
