+++ server/src/utils/asset.util.ts
@@ -5,11 +5,11 @@ import { UploadFieldName } from 'src/dtos/asset-media.dto';
 import { AuthDto } from 'src/dtos/auth.dto';
 import { AssetFileEntity } from 'src/entities/asset-files.entity';
 import { AssetFileType, AssetType, Permission } from 'src/enum';
-import { IEventRepository } from 'src/interfaces/event.interface';
 import { AuthRequest } from 'src/middleware/auth.guard';
 import { ImmichFile } from 'src/middleware/file-upload.interceptor';
 import { AccessRepository } from 'src/repositories/access.repository';
 import { AssetRepository } from 'src/repositories/asset.repository';
+import { EventRepository } from 'src/repositories/event.repository';
 import { PartnerRepository } from 'src/repositories/partner.repository';
 import { UploadFile } from 'src/services/asset-media.service';
 import { checkAccess } from 'src/utils/access';
@@ -139,7 +139,7 @@ export const getMyPartnerIds = async ({ userId, repository, timelineEnabled }: P
   return [...partnerIds];
 };
 
-export type AssetHookRepositories = { asset: AssetRepository; event: IEventRepository };
+export type AssetHookRepositories = { asset: AssetRepository; event: EventRepository };
 
 export const onBeforeLink = async (
   { asset: assetRepository, event: eventRepository }: AssetHookRepositories,
