+++ server/test/repositories/machine-learning.repository.mock.ts
@@ -1,7 +1,8 @@
-import { IMachineLearningRepository } from 'src/interfaces/machine-learning.interface';
+import { MachineLearningRepository } from 'src/repositories/machine-learning.repository';
+import { RepositoryInterface } from 'src/types';
 import { Mocked, vitest } from 'vitest';
 
-export const newMachineLearningRepositoryMock = (): Mocked<IMachineLearningRepository> => {
+export const newMachineLearningRepositoryMock = (): Mocked<RepositoryInterface<MachineLearningRepository>> => {
   return {
     encodeImage: vitest.fn(),
     encodeText: vitest.fn(),
